Stopped using the AHT10 files as a cFS shared library

+++ fsw/src/temp_app.h
@@ -69,7 +69,7 @@ typedef struct
     uint8 CmdCounter;
     uint8 ErrCounter;
 
-    uint8_t TimeCounter;
+    int TimeCounter;
 
     /*
     ** AHT10 data...
